Create whaleTalk.js
Whale Talk
Take a sentence like, 'Turpentine and turtles,' and translate it into its "whale talk" equivalent: 'UUEEIEE A UUEE'.

There are a few simple rules for translating text to whale speak:

There are no consonants. Only vowels.
The us and es are extra long, so we must double them in our program.
Once we have converted text to the whale speak, the result is sung slowly, as is a custom in the ocean.

To accomplish this translation, we can use our knowledge of loops. Let's get started!